backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
llama-cpp-python>=0.2.0
websockets>=12.0
